refactor: 💡 공감 포인트 모으기 시작 게이트웨이 위치 변경
